Fix business logic in slp-pkgmgr-server.
The current logic of pkgmgr-server is not robust to a burst of installation request.
This leads to have a deadlock into pkgmgr-server.

And it doens't answer to reuqest anymore

Force to wait backend end before returning for each job that is dependent of a backend

Make pm_queue thread safe. (we can have multiple request that leads to have concurencial meassge push )

BUG-Tizen=TC-2127

Change-Id: Ib3b3b4c4c9043b395869f0ee6c16da6db0da932d
Signed-off-by: Baptiste DURAND <baptiste.durand@open.eurogiciel.org>